fitness_tracker/requirements.txt
Django>=4.2

# REST framework and authentication
 djangorestframework
djongo
django-allauth
dj-rest-auth
djangorestframework-simplejwt
django-redis
mongoengine
pytest
pytest-django
python-dotenv 